+++ lib/sshkit.rb
@@ -6,15 +6,6 @@ module SSHKit
 
     attr_accessor :config
 
-    def capture_output(io, &block)
-      original_io = config.output
-      config.output = io
-      config.output.extend(SSHKit::Utils::CaptureOutputMethods)
-      yield
-    ensure
-      config.output = original_io
-    end
-
     def configure
       @@config ||= Configuration.new
       yield config