+++ test/unit/test_coordinator.rb
@@ -5,19 +5,15 @@ module SSHKit
 
   class TestCoordinator < UnitTest
 
-    CloseEnough = 0.01; # 10 msec
-
     def setup
       super
-      @s = String.new
+      @out = StringIO.new
+      SSHKit.config.output_verbosity = :debug
+      SSHKit.config.output = SSHKit::Formatter::SimpleText.new(@out)
       SSHKit.config.backend = SSHKit::Backend::Printer
     end
 
-    def tearddown
-      @s = nil
-    end
-
-    def block_to_run
+    def echo_time
       lambda do |host|
         execute "echo #{Time.now.to_f}"
       end
@@ -42,69 +38,66 @@ module SSHKit
     end
 
     def test_the_connection_manager_yields_the_host_to_each_connection_instance
-      spy = lambda do |host|
+      Coordinator.new(%w{1.example.com}).each do |host|
         execute "echo #{host.hostname}"
       end
-      String.new.tap do |str|
-        SSHKit.capture_output str do
-          Coordinator.new(%w{1.example.com}).each &spy
-        end
-        assert_equal "/usr/bin/env echo 1.example.com", str.strip
-      end
+      assert_equal "Command: echo 1.example.com\n", actual_output_commands.last
     end
 
     def test_the_connection_manaager_runs_things_in_parallel_by_default
-      SSHKit.capture_output @s do
-        Coordinator.new(%w{1.example.com 2.example.com}).each &block_to_run
-      end
-      assert_equal 2, results.length
-      assert_in_delta *results, CloseEnough
+      Coordinator.new(%w{1.example.com 2.example.com}).each &echo_time
+      assert_equal 2, actual_execution_times.length
+      assert_within_10_ms(actual_execution_times)
     end
 
     def test_the_connection_manager_can_run_things_in_sequence
-      SSHKit.capture_output @s do
-        Coordinator.new(%w{1.example.com 2.example.com}).each in: :sequence, &block_to_run
-      end
-      assert_equal 2, results.length
-      assert_operator (results.last - results.first), :>, 1.0
+      Coordinator.new(%w{1.example.com 2.example.com}).each in: :sequence, &echo_time
+      assert_equal 2, actual_execution_times.length
+      assert_at_least_1_sec_apart(actual_execution_times.first, actual_execution_times.last)
     end
 
     def test_the_connection_manager_can_run_things_in_sequence_with_wait
       start = Time.now
-      SSHKit.capture_output @s do
-        Coordinator.new(%w{1.example.com 2.example.com}).each in: :sequence, wait: 10, &block_to_run
-      end
+      Coordinator.new(%w{1.example.com 2.example.com}).each in: :sequence, wait: 10, &echo_time
       stop = Time.now
       assert_operator (stop - start), :>=, 10.0
     end
 
     def test_the_connection_manager_can_run_things_in_groups
-      SSHKit.capture_output @s do
-        Coordinator.new(
-          %w{
-            1.example.com
-            2.example.com
-            3.example.com
-            4.example.com
-            5.example.com
-            6.example.com
-          }
-        ).each in: :groups, &block_to_run
-      end
-      assert_equal 6, results.length
-      assert_in_delta *results[0..1], CloseEnough
-      assert_in_delta *results[2..3], CloseEnough
-      assert_in_delta *results[4..5], CloseEnough
-      assert_operator (results[2] - results[1]), :>, 1.0
-      assert_operator (results[4] - results[3]), :>, 1.0
+      Coordinator.new(
+        %w{
+          1.example.com
+          2.example.com
+          3.example.com
+          4.example.com
+          5.example.com
+          6.example.com
+        }
+      ).each in: :groups, &echo_time
+      assert_equal 6, actual_execution_times.length
+      assert_within_10_ms(actual_execution_times[0..1])
+      assert_within_10_ms(actual_execution_times[2..3])
+      assert_within_10_ms(actual_execution_times[4..5])
+      assert_at_least_1_sec_apart(actual_execution_times[1], actual_execution_times[2])
+      assert_at_least_1_sec_apart(actual_execution_times[3], actual_execution_times[4])
     end
 
     private
 
-    def results
-      @s.lines.collect do |line|
-        line.split(' ').last.to_f
-      end
+    def assert_at_least_1_sec_apart(first_time, last_time)
+      assert_operator (last_time - first_time), :>, 1.0
+    end
+
+    def assert_within_10_ms(array)
+      assert_in_delta *array, 0.01 # 10 msec
+    end
+
+    def actual_execution_times
+      actual_output_commands.map { |line| line.split(' ').last.to_f }
+    end
+
+    def actual_output_commands
+      @out.string.lines.select { |line| line.start_with?('Command:') }
     end
 
   end